requirements.txt
pynput
requests
python-telegram-bot
python-dotenv